move Pizza to models package
order it is not just a pizza
