+++ Task_2_2_1/src/main/java/ru/nsu/valikov/orders/models/Pizza.java
@@ -1,4 +1,4 @@
-package ru.nsu.valikov.orders.pizza;
+package ru.nsu.valikov.orders.models;
 
 import ru.nsu.valikov.orders.Order;
